web/requirements.txt
django==2.2.2
psycopg2-binary==2.8.3
uwsgi==2.0.18
requests==2.22.0
cloudinary==1.16.0
djangorestframework==3.9.4
django-cors-headers==3.0.2
django-cors-middleware==1.3.1
mercadopago==0.3.5
celery==4.3.0
redis==3.3.11
